separate out into exported object
